+++ Src/CSharpier.Rider/src/main/java/com/intellij/csharpier/CSharpierStartup.java
@@ -9,6 +9,7 @@ import org.jetbrains.annotations.NotNull;
 public class CSharpierStartup implements StartupActivity, DumbAware {
     @Override
     public void runActivity(@NotNull Project project) {
+        DotNetProvider.getInstance(project);
         CSharpierProcessProvider.getInstance(project);
         ApplicationManager.getApplication().getService(ReformatWithCSharpierOnSave.class);
     }